requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
anyio==3.7.1
arrow==1.2.3
async-timeout==4.0.3
attrs==23.1.0
CacheControl==0.13.1
cachetools==5.3.1
certifi==2023.7.22
cffi==1.15.1
charset-normalizer==3.2.0
click==8.1.6
colorama==0.4.6
cryptography==41.0.3
dataclasses-json==0.5.14
demjson3==3.0.6
fastapi==0.101.1
firebase-admin==6.2.0
frozenlist==1.4.0
google-api-core==2.11.1
google-api-python-client==2.96.0
google-auth==2.22.0
google-auth-httplib2==0.1.0
google-cloud-core==2.3.3
google-cloud-firestore==2.11.1
google-cloud-storage==2.10.0
google-crc32c==1.5.0
google-resumable-media==2.5.0
googleapis-common-protos==1.60.0
greenlet==2.0.2
grpcio==1.57.0
grpcio-status==1.57.0
h11==0.14.0
httplib2==0.22.0
idna==3.4
langchain==0.0.264
langsmith==0.0.22
marshmallow==3.20.1
msgpack==1.0.5
multidict==6.0.4
mypy-extensions==1.0.0
nanoid==2.0.0
numexpr==2.8.5
numpy==1.25.2
oauth2==1.9.0.post1
openai==0.27.8
openapi-schema-pydantic==1.2.4
packaging==23.1
proto-plus==1.22.3
protobuf==4.24.0
pyasn1==0.5.0
pyasn1-modules==0.3.0
pycparser==2.21
pydantic==1.10.12
PyJWT==2.8.0
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3
PyYAML==6.0.1
requests==2.31.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
SQLAlchemy==2.0.19
starlette==0.27.0
tenacity==8.2.3
tqdm==4.66.1
typing-inspect==0.9.0
typing_extensions==4.7.1
uritemplate==4.1.1
urllib3==1.26.16
uvicorn==0.23.2
yarl==1.9.2
